refactor :: moved some reusable functions into helpers.js and added some lightweight documentation to existing functions
