+++ package.json
@@ -13,15 +13,15 @@
   "license": "ISC",
   "devDependencies": {
     "@11ty/eleventy": "^0.12.1",
-    "alpinejs": "^2.8.1",
+    "alpinejs": "^2.8.2",
     "autoprefixer": "^10.2.5",
     "cross-env": "^7.0.3",
     "postcss-cli": "^8.3.1",
-    "tailwindcss": "^2.0.4"
+    "tailwindcss": "^2.1.1"
   },
   "dependencies": {
     "luxon": "^1.26.0",
-    "markdown-it": "^12.0.4",
+    "markdown-it": "^12.0.6",
     "markdown-it-anchor": "^7.1.0",
     "markdown-it-footnote": "^3.0.2"
   }
